requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
faiss-cpu
python-dotenv
pymupdf
fastembed